[meta] Rename cdsl/inst to cdsl/instructions;
